lib: obstacles: move class Polygon to its own file
